requirements.txt
opencv-python
transformers
torch
safetensors
mediapipe
gdown
captum
pillow